Working Solution Found!
Finally created a modular block of code that correctly counts the number of syllables in any given piece of text.

The code works using the syllapy package. Essentially, it separates the text into individual words, then uses the  'syllapy._syllables' function for each word.  After a word is checked for it's number of syllables, it then initializes a variable to an integer which represents the number of syllables in that word.  This integer is then appended to a list of numbers, each of which representing a different word. finally, after all words have been checked and converted to list integers, the sum function is used to add all those integers and return the number of syllables in the text.